refactor: update recent watched tests to use Flow
