+++ data/src/test/java/com/london/data/datasource/local/recent/watched/RecentWatchedMoviesDataSourceTest.kt
@@ -1,36 +1,37 @@
 package com.london.data.datasource.local.recent.watched
 
 import com.google.common.truth.Truth.assertThat
-import com.london.data.local.database.dao.recent.whatched.movie.RecentWatchedMoviesDao
+import com.london.data.local.database.dao.recent.watched.movie.RecentWatchedMoviesDao
 import com.london.data.local.model.recent.watched.RecentWatchedMovieLocal
+import com.london.data.local.source.recent.watched.RecentWatchedMoviesDataSource
 import io.mockk.Runs
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.just
 import io.mockk.mockk
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.flow.single
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
 import org.junit.Test
 
 class RecentWatchedMoviesDataSourceTest {
     private lateinit var recentWatchedMoviesDao: RecentWatchedMoviesDao
-    private lateinit var recentWatchedMoviesDataSource: com.london.data.local.source.recent.watched.RecentWatchedMoviesDataSource
+    private lateinit var recentWatchedMoviesDataSource: RecentWatchedMoviesDataSource
 
     @Before
     fun setUp() {
         recentWatchedMoviesDao = mockk()
-        recentWatchedMoviesDataSource =
-            com.london.data.local.source.recent.watched.RecentWatchedMoviesDataSource(
-                recentWatchedMoviesDao
-            )
+        recentWatchedMoviesDataSource = RecentWatchedMoviesDataSource(recentWatchedMoviesDao)
     }
 
     @Test
     fun `getAll should return list of RecentWatchedTvShowLocal when dao returns list`() = runTest {
         // Given
-        coEvery { recentWatchedMoviesDao.getAll() } returns recentWatchedMovieLocalList
+        coEvery { recentWatchedMoviesDao.getAll() } returns flowOf(recentWatchedMovieLocalList)
         // When
-        val result = recentWatchedMoviesDataSource.getAll()
+        val result = recentWatchedMoviesDataSource.getAll().single()
         // Then
         assertThat(result).isEqualTo(recentWatchedMovieLocalList)
     }
@@ -38,9 +39,9 @@ class RecentWatchedMoviesDataSourceTest {
     @Test
     fun `getAll should return empty list when dao returns empty list`() = runTest {
         // Given
-        coEvery { recentWatchedMoviesDao.getAll() } returns emptyList()
+        coEvery { recentWatchedMoviesDao.getAll() } returns flowOf(emptyList())
         // When
-        val result = recentWatchedMoviesDataSource.getAll()
+        val result = recentWatchedMoviesDataSource.getAll().single()
         // Then
         assertThat(result).isEmpty()
     }
@@ -58,9 +59,9 @@ class RecentWatchedMoviesDataSourceTest {
     @Test
     fun `getAll should return empty list when dao throws exception`() = runTest {
         // Given
-        coEvery { recentWatchedMoviesDao.getAll() } throws Exception()
+        coEvery { recentWatchedMoviesDao.getAll() } returns flow { throw Exception("DAO error") }
         // When
-        val result = recentWatchedMoviesDataSource.getAll()
+        val result = recentWatchedMoviesDataSource.getAll().single()
         // Then
         assertThat(result).isEmpty()
     }
@@ -70,5 +71,4 @@ class RecentWatchedMoviesDataSourceTest {
         val recentWatchedMovieLocalList =
             listOf(recentWatchedMovieLocal, recentWatchedMovieLocal, recentWatchedMovieLocal)
     }
-
 }
