+++ data/src/test/java/com/london/data/datasource/local/recent/watched/RecentWatchedTvShowsDataSourceTest.kt
@@ -1,36 +1,37 @@
 package com.london.data.datasource.local.recent.watched
 
 import com.google.common.truth.Truth.assertThat
-import com.london.data.local.database.dao.recent.whatched.tvshow.RecentWatchedTvShowsDao
+import com.london.data.local.database.dao.recent.watched.tvshow.RecentWatchedTvShowsDao
 import com.london.data.local.model.recent.watched.RecentWatchedTvShowLocal
+import com.london.data.local.source.recent.watched.RecentWatchedTvShowsDataSource
 import io.mockk.Runs
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.just
 import io.mockk.mockk
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.flow.single
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
 import kotlin.test.Test
 
 class RecentWatchedTvShowsDataSourceTest {
     private lateinit var recentWatchedTvShowsDao: RecentWatchedTvShowsDao
-    private lateinit var recentWatchedTvShowsDataSource: com.london.data.local.source.recent.watched.RecentWatchedTvShowsDataSource
+    private lateinit var recentWatchedTvShowsDataSource: RecentWatchedTvShowsDataSource
 
     @Before
     fun setUp() {
         recentWatchedTvShowsDao = mockk()
-        recentWatchedTvShowsDataSource =
-            com.london.data.local.source.recent.watched.RecentWatchedTvShowsDataSource(
-                recentWatchedTvShowsDao
-            )
+        recentWatchedTvShowsDataSource = RecentWatchedTvShowsDataSource(recentWatchedTvShowsDao)
     }
 
     @Test
     fun `getAll should return list of RecentWatchedTvShowLocal when dao returns list`() = runTest {
         // Given
-        coEvery { recentWatchedTvShowsDao.getAll() } returns recentWatchedTvShowLocalList
+        coEvery { recentWatchedTvShowsDao.getAll() } returns flowOf(recentWatchedTvShowLocalList)
         // When
-        val result = recentWatchedTvShowsDataSource.getAll()
+        val result = recentWatchedTvShowsDataSource.getAll().single()
         // Then
         assertThat(result).isEqualTo(recentWatchedTvShowLocalList)
     }
@@ -38,9 +39,9 @@ class RecentWatchedTvShowsDataSourceTest {
     @Test
     fun `getAll should return empty list when dao returns empty list`() = runTest {
         // Given
-        coEvery { recentWatchedTvShowsDao.getAll() } returns emptyList()
+        coEvery { recentWatchedTvShowsDao.getAll() } returns flowOf(emptyList())
         // When
-        val result = recentWatchedTvShowsDataSource.getAll()
+        val result = recentWatchedTvShowsDataSource.getAll().single()
         // Then
         assertThat(result).isEmpty()
     }
@@ -58,9 +59,9 @@ class RecentWatchedTvShowsDataSourceTest {
     @Test
     fun `getAll should return empty list when dao throws exception`() = runTest {
         // Given
-        coEvery { recentWatchedTvShowsDao.getAll() } throws Exception()
+        coEvery { recentWatchedTvShowsDao.getAll() } returns flow { throw Exception("DAO error") }
         // When
-        val result = recentWatchedTvShowsDataSource.getAll()
+        val result = recentWatchedTvShowsDataSource.getAll().single()
         // Then
         assertThat(result).isEmpty()
     }
